Student Guide Sections Added
Added in labels, descriptions, and images for student guide sections. Also included download buttons for the written versions of the guide. Files added include pdf & docx versions of the guide and images of the model. Links for all images, links, and downloads will be added next once downloadable files are on GitHub (this commit) and  additional html files are created (guide pages).